requirements.txt
# ─────────────────────────────────────────────
# AI Banking Chatbot - Requirements
# Compatible with Python 3.14 on Mac
# Install: pip install -r requirements.txt
# ─────────────────────────────────────────────

# ── Core LangChain ────────────────────────────
langchain>=1.0.0
langchain-community>=0.3.0
langchain-huggingface>=0.1.0
langchain-chroma>=0.1.4
langchain-groq>=0.2.0

# ── HuggingFace (embeddings only — runs locally) ──
huggingface-hub>=0.24.7
sentence-transformers>=3.1.1
transformers>=4.44.2
tokenizers>=0.19.1

# ── Vector Database ───────────────────────────
chromadb>=0.5.0

# ── PDF Loading ───────────────────────────────
pypdf>=4.3.1

# ── Streamlit UI ──────────────────────────────
streamlit>=1.39.0

# ── Environment Variables ─────────────────────
python-dotenv>=1.0.1

# ── Core dependencies ─────────────────────────
pydantic>=2.9.2
numpy>=2.0.0
torch>=2.9.0